added blue squares, approximating their color
this one is hard. i used the standard blending equation to help me determine what the original RGB values were. it was a bit of a guessing game though because the equation requires an alpha value, and I could only guess.

i recorded the results of this trial and error and went with that - in the end I probably could have just guessed via the console.